REFACTOR: Add test coverage for current user.
Looking to add better test coverage for the current user on the spec helper.
