+++ app/helpers/application_helper.rb
@@ -1,2 +1,6 @@
 module ApplicationHelper
+
+  def log_in(user)
+    session[:user_id] = user.id
+  end
 end
