+++ app/helpers/sessions_helper.rb
@@ -5,8 +5,4 @@ module SessionsHelper
       @current_user ||= User.find_by(id: session[:user_id])
     end
   end
-
-  def log_in(user)
-    session[:user_id] = user.id
-  end
 end
