+++ spec/helpers/application_helper_spec.rb
@@ -1,0 +1,11 @@
+require 'rails_helper'
+
+RSpec.describe ApplicationHelper do
+  describe "#log_in" do
+    it "creates a session id" do
+      user = double("User", :id => 123)
+      log_in(user)
+      expect(session[:user_id]).to eq(123)
+    end
+  end
+end
